ref(seer): Rename useAgentHooks to seerAgentHooks
This thing will have more variety inside it in a next step
