+++ static/gsApp/views/seerAutomation/components/seerAgentHooks.tsx
